refactor: move bom writing functionality to separate class
This functionality might be useful for anybody that writes a SBOM, so
move it into its own class.
