+++ src/debsbom/bomwriter/__init__.py
@@ -1,0 +1,5 @@
+# Copyright (C) 2025 Siemens
+#
+# SPDX-License-Identifier: MIT
+
+from .bomwriter import BomWriter
